(webdriverio): resolve PR feedback
