+++ packages/webdriverio/src/commands/element/isStable.ts
@@ -1,4 +1,6 @@
+import { ELEMENT_KEY } from '../../constants.js'
 import { getBrowserObject } from '../../utils/index.js'
+import isElementStable from '../../scripts/isElementStable.js'
 
 /**
  *
@@ -7,37 +9,38 @@ import { getBrowserObject } from '../../utils/index.js'
  * <example>
     :index.html
     <head>
-    <style>
-        div {
-        width: 200px;
-        height: 200px;
-        background-color: red;
-        animation: 3s 0s alternate slidein;
-        }
-        @keyframes slidein {
-        from {
-            margin-left: 100%;
-            width: 300%;
-        }
+        <style>
+            #has-animation {
+                width: 200px;
+                height: 200px;
+                background-color: red;
+                animation: 3s 0s alternate slidein;
+            }
+            @keyframes slidein {
+                from {
+                    margin-left: 100%;
+                    width: 300%;
+                }
 
-        to {
-            margin-left: 0%;
-            width: 100%;
-        }
-        }
-    </style>
+                to {
+                    margin-left: 0%;
+                    width: 100%;
+                }
+            }
+        </style>
     </head>
     <body>
-    <div></div>
+        <div #has-animation></div>
+        <div #has-no-animation></div>
     </body>
 
     :isStable.js
     it('should detect if an element is stable', async () => {
         let element = await $('#has-animation');
-        console.log(await element.isStable()); // outputs: true
+        console.log(await element.isStable()); // outputs: false
 
         element = await $('#has-no-animation')
-        console.log(await element.isStable()); // outputs: false
+        console.log(await element.isStable()); // outputs: true
     });
  * </example>
  *
@@ -48,25 +51,8 @@ import { getBrowserObject } from '../../utils/index.js'
  */
 export async function isStable (this: WebdriverIO.Element) {
     const browser = getBrowserObject(this)
-    return await browser.executeAsync((elem, done) => {
-        if (document.visibilityState === 'hidden') {
-            throw Error('You are are checking for animations on an inactive tab, animations do not run for inactive tabs')
-        }
-        try {
-            const previousPosition = elem.getBoundingClientRect()
-            requestAnimationFrame(() => {
-                requestAnimationFrame(() => {
-                    const currentPosition = elem.getBoundingClientRect()
-                    for (const prop in previousPosition) {
-                        if (previousPosition[(prop as keyof DOMRect)] !== currentPosition[(prop as keyof DOMRect)]) {
-                            done(false)
-                        }
-                    }
-                    done(true)
-                })
-            })
-        } catch (error) {
-            done(false)
-        }
-    }, await this as unknown as HTMLElement)
+    return await browser.executeAsync(isElementStable, {
+        [ELEMENT_KEY]: this.elementId, // w3c compatible
+        ELEMENT: this.elementId // jsonwp compatible
+    } as any as HTMLElement)
 }
